ViewModelの実装をViewModelImplに変更 #1
